Mejorar y abstraer la gestión de los sensores

+++ src/drivers/DHT11HumAndTemp.h
@@ -4,7 +4,7 @@
 
 template <class T>
 struct DHTResult {
-    uint16_t error;
+    int error;
     T result;
 };
